requirements.txt
numpy<2
scipy
opencv-python
gdown
pillow
gradio
sklearn